pyproject.toml
[tool.poetry]
name = "ltc-django-boilerplate"
version = "0.1.0"
description = "A production-ready Django boilerplate with Poetry, Docker, and modern best practices."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}
django-environ = "^0.11.2"
gunicorn = "^22.0.0"
whitenoise = "^6.6.0"
python-json-logger = "^2.0.7"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
pre-commit = "^3.7.0"
pytest = "^8.2.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C90"]
ignore = []

[tool.ruff.format]
quote-style = "double"
